ai_personal_assistant/requirements.txt
# Core AI and Voice Processing
SpeechRecognition==3.10.0
pyttsx3==2.90
PyAudio==0.2.13

# Office Automation
python-pptx==0.6.23
pywin32==306; sys_platform == 'win32'
pyobjc-framework-Cocoa==10.1; sys_platform == 'darwin'
pyobjc-framework-ScriptingBridge==10.1; sys_platform == 'darwin'

# Communication & Messaging
pywhatkit==5.4
selenium==4.16.0
webdriver-manager==4.0.1

# Calendar and Scheduling
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0

# Natural Language Processing
openai==1.6.1
anthropic==0.7.8

# Logging and Debugging
colorama==0.4.6
rich==13.7.0
